Stabilize archive object telemetry

+++ apps/backend/src/history-scan-coordinator/domain/history-archive-object/HistoryArchiveObject.ts
@@ -67,7 +67,7 @@ export class HistoryArchiveObject extends CoreEntity {
 	@Column('text', { nullable: true })
 	public bucketHash!: string | null;
 
-	@Column('integer', { nullable: true })
+	@Column('bigint', { nullable: true })
 	public bytesDownloaded!: number | null;
 
 	@Column('integer', { default: 0 })
